Add files via upload
The back-end resources for the P300 have been updated, with the MATLAB scripts altered to include a better signal processing scheme for the P300. (Using the low-frequency Dual Lagrange method, called the DALERP paradigm in BCILAB). Additionally, some extra functionality has been included in the row-column implementation of the P300 tool.