Add files via upload
Updated the P300 processing scripts of 'test_controller', 'model_setup' to better support the ParadigmWindowmeans from BCILAB. Included some optional flags to visualize ROC curves from 'test_controller', and included a default LDA analysis of the P300 as the ML approach. Also included two scripts - 'load_test_controller' to let the user input parameters of interest, and a work-in-progress script for grid-search optimization of window parameters. (Optimization is a work in progress - currently just doing a slow nested for-loop search, to be updated at a later date).